requirements.txt
streamlit==1.28.1
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0